Changed the PageMedia object to a Media object
Rather than a one to one relationship between a page and its media, now
we have a 1 to many relationship. changed all the code and tests to
reflect this.

This might pose some problems, so the decision is not final. Just wanted
to commit In case my PC blows up.